use standard __git_ps1 helper in prompt
We had rolled our own helper function for generating the git prompt. git
comes packaged with these helpers built and maintained by the community.
The __git_ps1 helper is massively configurable to support a load of
different useful features. For now I've enabled the show_dirty_state
flag so that the status of the repo is shown in the prompt.